requirements/local.txt
-r base.txt

# Dev server & debugging
django-debug-toolbar==4.4.6
Werkzeug==3.1.3
ipython==8.29.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
Faker==33.1.0

# Linting / formatting (also used in CI)
ruff==0.8.4
pre-commit==4.0.1
